Remove SocketInterface classe
