+++ core/Makefile
@@ -38,7 +38,6 @@ OBJS_KSERVER = core/kserver.o              \
                core/crypto/base64.o        \
                core/crypto/sha1.o          \
                core/kserver_syslog.o       \
-               core/socket_interface.o     \
                core/signal_handler.o       \
                core/binary_parser.o
